# services/orca-core/src/main/resources/openapi/orca-core.yaml
openapi: 3.0.3
...
info:
  title: orca-core
  description: |
    The world as configured (§C1) — tenancy, sites and lanes, users, devices, design artifacts, custom entities, licence verification.
  version: 0.0.1
servers:
  - url: http://localhost:8081
    description: Local development
...
paths:
...
  /api/v1/lanes/{laneExternalId}/devices:
    get:
      tags: [devices]
      operationId: listLaneDevices
      summary: The device inventory of a lane, port layout and presets included
      parameters:
        - name: laneExternalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The devices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevicesEnvelope"
        "401":
          description: No credential, or one that did not verify
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "404":
          description: No such lane
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
    post:
      tags: [devices]
      operationId: registerDevice
      summary: Register a device at a lane
      description: |
        The type is a catalog code — the vocabulary is the seeded catalog, not a
        free string. No credential field exists on this surface: device-credential
        storage is a proposed design awaiting the product owner (WP3 report).
      parameters:
        - name: laneExternalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDeviceRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceEnvelope"
        "401":
          description: No credential, or one that did not verify
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "404":
          description: No such lane
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "422":
          description: The device type code is not in the catalog
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
  /api/v1/devices/{deviceExternalId}:
    patch:
      tags: [devices]
      operationId: updateDevice
      summary: Update a device
      parameters:
        - name: deviceExternalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeviceRequest"
      responses:
        "200":
          description: The device after the change
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceEnvelope"
        "401":
          description: No credential, or one that did not verify
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "404":
          description: No such device
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "422":
          description: The device type code is not in the catalog
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
    delete:
      tags: [devices]
      operationId: retireDevice
      summary: Retire a device
      description: Retirement, not deletion (§D3) — the device leaves every published view and its history stays.
      parameters:
        - name: deviceExternalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The retired device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceEnvelope"
        "401":
          description: No credential, or one that did not verify
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "404":
          description: No such device
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
  /api/v1/devices/{deviceExternalId}/io-assignments:
    put:
      tags: [devices]
      operationId: replaceIoAssignments
      summary: Replace the device's IO port layout
      description: Declarative — the request's layout is the whole layout.
      parameters:
        - name: deviceExternalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/IoAssignmentItem"
      responses:
        "200":
          description: The device with its new layout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceEnvelope"
        "400":
          description: The request repeats a port
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "401":
          description: No credential, or one that did not verify
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "404":
          description: No such device
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "422":
          description: A port name code is not in the catalog
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
  /api/v1/devices/{deviceExternalId}/perspectives:
    put:
      tags: [devices]
      operationId: replacePerspectives
      summary: Replace the device's named PTZ presets
      description: |
        Preset STORAGE only. Execution is a direct camera call and belongs to
        the hardware boundary (register NEW-5, narrowed 7 Aug 2026).
      parameters:
        - name: deviceExternalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/PerspectiveItem"
      responses:
        "200":
          description: The device with its new presets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceEnvelope"
        "400":
          description: The request repeats a preset name
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "401":
          description: No credential, or one that did not verify
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "404":
          description: No such device
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
...
components:
  schemas:
...
    DeviceMode:
      type: string
      description: 1.x compared this with LOWER(?) and leaned on collation; here it is one casing.
      enum: [IO, DATA_CAPTURE]
    PortType:
      type: string
      description: The unified port-type enum — TitleCase in one 1.x table, lowercase in another; one casing here.
      enum: [INPUT, OUTPUT, RELAY, AUDIO, TONE]
...
    IoAssignmentItem:
      type: object
      required: [portType, ioPort]
      properties:
        portType:
          $ref: "#/components/schemas/PortType"
        ioPort:
          type: integer
          format: int32
        portNameCode:
          type: string
          description: A code from the IO port name catalog (DI1 … DO3 … R2 … FRONT_MIC).
        portAliasName:
          type: string
          maxLength: 200
        reverseState:
          type: boolean
          default: false
        initialHigh:
          type: boolean
          default: false
        dataCapture:
          type: boolean
          default: false
        gosAudio:
          type: boolean
          default: false
        produceTrueMessage:
          type: boolean
          default: true
          description: Edge-suppression semantics — edge reads these; silence is configured, never accidental.
        produceFalseMessage:
          type: boolean
          default: true
        waitTime:
          type: integer
          format: int32
    PerspectiveItem:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 200
        pan:
          type: number
        tilt:
          type: number
        zoom:
          type: number
    DeviceSummary:
      type: object
      required: [externalId, laneExternalId, siteExternalId, deviceTypeCode, name, protocol,
                 ioAssignments, perspectives, retired]
      properties:
        externalId:
          type: string
        laneExternalId:
          type: string
        siteExternalId:
          type: string
        deviceTypeCode:
          type: string
        name:
          type: string
        address:
          type: string
        mode:
          $ref: "#/components/schemas/DeviceMode"
        alias:
          type: string
        model:
          type: string
        firmwareVersion:
          type: string
        description:
          type: string
        resolution:
          type: string
        ipAddress:
          type: string
        streamType:
          type: string
        port:
          type: integer
          format: int32
        protocol:
          type: string
        manufacturer:
          type: string
        deviceUrl:
          type: string
        devicePortalUrl:
          type: string
        assemblyName:
          type: string
          description: .NET plugin-load parameter — load-bearing for the frozen config poll (§D2).
        className:
          type: string
        dataCaptureMode:
          type: string
        waitTime:
          type: integer
          format: int32
        ioAssignments:
          type: array
          items:
            $ref: "#/components/schemas/IoAssignmentItem"
        perspectives:
          type: array
          items:
            $ref: "#/components/schemas/PerspectiveItem"
        retired:
          type: boolean
        createdAt:
          type: string
          format: date-time
    RegisterDeviceRequest:
      type: object
      required: [name, deviceTypeCode]
      properties:
        name:
          type: string
          maxLength: 200
        deviceTypeCode:
          type: string
        address:
          type: string
          maxLength: 512
        mode:
          $ref: "#/components/schemas/DeviceMode"
        alias:
          type: string
          maxLength: 200
        model:
          type: string
          maxLength: 200
        firmwareVersion:
          type: string
          maxLength: 20
        description:
          type: string
          maxLength: 500
        resolution:
          type: string
          maxLength: 50
        ipAddress:
          type: string
          maxLength: 45
        streamType:
          type: string
          maxLength: 50
        port:
          type: integer
          format: int32
        protocol:
          type: string
          maxLength: 10
        manufacturer:
          type: string
          maxLength: 200
        deviceUrl:
          type: string
          maxLength: 512
        devicePortalUrl:
          type: string
          maxLength: 512
        assemblyName:
          type: string
          maxLength: 200
        className:
          type: string
          maxLength: 255
        dataCaptureMode:
          type: string
          maxLength: 25
        waitTime:
          type: integer
          format: int32
    UpdateDeviceRequest:
      type: object
      description: Absent fields are unchanged.
      properties:
        name:
          type: string
          maxLength: 200
        deviceTypeCode:
          type: string
        address:
          type: string
          maxLength: 512
        mode:
          $ref: "#/components/schemas/DeviceMode"
        alias:
          type: string
          maxLength: 200
        model:
          type: string
          maxLength: 200
        firmwareVersion:
          type: string
          maxLength: 20
        description:
          type: string
          maxLength: 500
        resolution:
          type: string
          maxLength: 50
        ipAddress:
          type: string
          maxLength: 45
        streamType:
          type: string
          maxLength: 50
        port:
          type: integer
          format: int32
        protocol:
          type: string
          maxLength: 10
        manufacturer:
          type: string
          maxLength: 200
        deviceUrl:
          type: string
          maxLength: 512
        devicePortalUrl:
          type: string
          maxLength: 512
        assemblyName:
          type: string
          maxLength: 200
        className:
          type: string
          maxLength: 255
        dataCaptureMode:
          type: string
          maxLength: 25
        waitTime:
          type: integer
          format: int32
    DeviceEnvelope:
      allOf:
        - $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/DeviceSummary"
    DevicesEnvelope:
      allOf:
        - $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/DeviceSummary"
...